+++ seedfarmer/commands/_module_commands.py
@@ -175,7 +175,7 @@ def destroy_module(
     remove_ssm = [f"seedfarmer remove moduledata -d {deployment_name} -g {group_name} -m {module_manifest_name}"]
 
     export_info = [
-        f"export DEPLOYEMNT={deployment_name}",
+        f"export DEPLOYMENT={deployment_name}",
         f"export GROUP={group_name}",
         f"export MODULE={module_manifest_name}",
     ]