+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### Fixes
 - exit code on on failed modules changed from 0 1o 1
+- changed typo `Deployemnt` to `Deployment` everywhere
 
 
 ## v2.1.0 (2022-10-03)